Refactors data layer to use Long IDs
Migrates Deck and Flashcard IDs from Integer to Long to accommodate larger datasets.

Introduces EntityFactory objects for converting between data entities and domain entities.

Adds OpenAI integration for generating flashcard decks.

+++ android/app/src/main/java/com/synergyboat/flashcardAi/data/dao/FlashcardDao.kt
@@ -11,7 +11,7 @@ interface FlashcardDao {
     suspend fun createFlashcard(flashcard: FlashcardEntity)
 
     @Query("SELECT * FROM flashcard WHERE deckId = :deckId")
-    suspend fun getAllFlashcardsFromDeckId(deckId: Int): List<FlashcardEntity>
+    suspend fun getAllFlashcardsFromDeckId(deckId: Long): List<FlashcardEntity>
 
     @Update
     suspend fun updateFlashcard(flashcard: FlashcardEntity)
@@ -20,5 +20,5 @@ interface FlashcardDao {
     suspend fun deleteFlashcard(flashcard: FlashcardEntity)
 
     @Query("DELETE FROM flashcard WHERE id = :flashcardId")
-    suspend fun deleteFlashcardById(flashcardId: Int)
+    suspend fun deleteFlashcardById(flashcardId: Long)
 }
