+++ android/app/src/main/java/com/synergyboat/flashcardAi/domain/repository/FlashcardRepository.kt
@@ -4,8 +4,8 @@ import com.synergyboat.flashcardAi.domain.entities.Flashcard
 
 interface FlashcardRepository {
     suspend fun createFlashcard(flashcard: Flashcard): Flashcard
-    suspend fun getFlashcardsByDeck(deckId: Int): List<Flashcard>
-    suspend fun deleteFlashcard(id: Int)
+    suspend fun getFlashcardsByDeck(deckId: Long): List<Flashcard>
+    suspend fun deleteFlashcard(id: Long)
     suspend fun updateFlashcard(flashcard: Flashcard): Flashcard
-    suspend fun getFlashcardById(id: Int): Flashcard?
+    suspend fun getFlashcardById(id: Long): Flashcard?
 }
