+++ android/app/src/main/java/com/synergyboat/flashcardAi/domain/repository/ai/AiGeneratorRepository.kt
@@ -1,5 +1,5 @@
 package com.synergyboat.flashcardAi.domain.repository.ai
 
 interface AiGeneratorRepository {
-    suspend fun generateDeck(deckId: Int?, prompt: String, count: Int): String
+    suspend fun generateDeck(deckId: Long?, prompt: String, count: Int): String
 }
