+++ android/app/src/main/java/com/synergyboat/flashcardAi/domain/usecase/DeleteDeckUseCase.kt
@@ -5,7 +5,7 @@ import com.synergyboat.flashcardAi.domain.repository.DeckRepository
 class DeleteDeckUseCase(
     private val repository: DeckRepository
 ) {
-    suspend operator fun invoke(deckId: Int) {
+    suspend operator fun invoke(deckId: Long) {
         repository.deleteDeck(deckId)
     }
 }
